+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## UNRELEASED
 
-- fixed: instance reset default template now falling back to current instance template
+## Bug fixes
+
+- compute: instance reset default template now falling back to current instance template
+- compute: remmove uri and tlssni fields when nlb service healthcheck is "tcp"
 
 ## 1.72.0